DB seed data, docs, and schema validation fixes
- Seed: idempotent reference data (jurisdictions, provinces, tax years, income/
  expense/asset/liability/document types, operators) + a full worked example of
  the rules-as-data model (Medical Expense Credit 2025: fact defs, constants,
  RPN formula + input bindings, published rule version, AND condition tree, and
  an apply_credit outcome), with notes on append-only 2026 versioning.
- README (Alembic revision map + run/maintenance guide) and DATA_DICTIONARY.
- Fixes found by executing the full migration on PostgreSQL 16:
  * UUIDv7 generator: cast bigint bit-ops to int for set_byte().
  * updated_at trigger loop: use pg_catalog and skip partition children
    (they inherit the trigger from the partitioned parent).

Validated end-to-end: all 20 files apply clean; 87 tables + 3 partitioned
parents across 14 schemas; UUIDv7 time-ordered; audit trigger + append-only
enforcement working; RLS isolates each user to their own rows.

+++ backend/db/sql/00_extensions_roles.sql
@@ -48,13 +48,13 @@ DECLARE
     rand_bytes bytea := gen_random_bytes(10);
     uuid_bytes bytea;
 BEGIN
-    -- 48 bits unix ms timestamp
-    uuid_bytes := set_byte('\x00000000000000000000000000000000'::bytea, 0, (unix_ms >> 40) & 255);
-    uuid_bytes := set_byte(uuid_bytes, 1, (unix_ms >> 32) & 255);
-    uuid_bytes := set_byte(uuid_bytes, 2, (unix_ms >> 24) & 255);
-    uuid_bytes := set_byte(uuid_bytes, 3, (unix_ms >> 16) & 255);
-    uuid_bytes := set_byte(uuid_bytes, 4, (unix_ms >> 8) & 255);
-    uuid_bytes := set_byte(uuid_bytes, 5, unix_ms & 255);
+    -- 48 bits unix ms timestamp (cast to int; set_byte's value arg is integer)
+    uuid_bytes := set_byte('\x00000000000000000000000000000000'::bytea, 0, ((unix_ms >> 40) & 255)::int);
+    uuid_bytes := set_byte(uuid_bytes, 1, ((unix_ms >> 32) & 255)::int);
+    uuid_bytes := set_byte(uuid_bytes, 2, ((unix_ms >> 24) & 255)::int);
+    uuid_bytes := set_byte(uuid_bytes, 3, ((unix_ms >> 16) & 255)::int);
+    uuid_bytes := set_byte(uuid_bytes, 4, ((unix_ms >> 8) & 255)::int);
+    uuid_bytes := set_byte(uuid_bytes, 5, (unix_ms & 255)::int);
     -- 74 bits randomness
     uuid_bytes := set_byte(uuid_bytes, 6, get_byte(rand_bytes, 0));
     uuid_bytes := set_byte(uuid_bytes, 7, get_byte(rand_bytes, 1));
